CMakeLists.txt: don't check for signals Boost library
:Release Notes:
Deprecated signals library was dropped in 1.69.0 release:
https://www.boost.org/users/history/version_1_69_0.html

Signals (v1) is now removed. Its removal was announced in 1.68 and its
deprecation was announced in 1.54. Boost 1.68 is the last release that
provides this library. Users are encouraged to use Signals2 instead. The
Boost community thanks Douglas Gregor for his work on Signals which
served its users well and which also inspired Signals2.

This component is already using Signals2 so just drop the check for old
unused component.

This is Copied from http://gpro.lge.com/239698

:Detailed Notes:
Fixes:
http://caprica.lgsvl.com:8080/Errors/Details/1437538

:Testing Performed:

:QA Notes:

:Issues Addressed:
[PLAT-78370] sam, notificationmgr, appinstalld, sensord: are using long

Change-Id: Ia4d634c2386c8ed09e4ce7b9d89fd21448a3dd03